+++ scripts/bash/reset.sh
@@ -1,12 +1,6 @@
 #!/bin/bash
 # Function to unset internal variables and avoid script errors
 function reset () {
-	if [[ "${@}" ]]; then
-		unset FILE1;
-		for FILE1 in "${@}"; do
-			unset $FILE1
-		done
-	else
 		unset ARG;
 		unset command;
 		unset diagnose;
@@ -18,5 +12,5 @@ function reset () {
 		unset missing;
 		unset status;
 		unset variable;
-	fi
+		unset this_pkg
 }